+++ src/language/fsharp.rs
@@ -201,6 +201,7 @@ static DEFINITION: LanguageDef = LanguageDef {
     container_body_kinds: &[],
     extract_container_name: Some(extract_container_name_fsharp),
     extract_qualified_method: None,
+    post_process_chunk: None,
 };
 
 pub fn definition() -> &'static LanguageDef {